1_reverse + py

+++ check/checker.sh
@@ -23,7 +23,12 @@ TASK_ID=$3
 cp ../problems/$TASK_ID/input.txt ./input.txt
 cp ../problems/$TASK_ID/output.txt ./sample.txt
 
-"${!LANG}" ../solutions/$TASK_ID/$TEXT -o runnable
-./runnable < input.txt  > output.txt 
+if [[ ! -z ${compile[$LANG]} ]] ; then
+	"${!LANG}" ../solutions/$TASK_ID/$TEXT -o runnable
+	./runnable < input.txt  > output.txt 
+	rm runnable
+else
+	../solutions/$TASK_ID/$TEXT < input.txt  > output.txt 
+fi
+
 diff -q sample.txt output.txt
-rm runnable
